+++ internal/provider/bounce_settings_data_source_test.go
@@ -1,0 +1,32 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestAccBounceSettingsDataSource(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Read testing
+			{
+				Config: testAccBounceSettingsDataSourceConfig,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet("data.sendgrid_bounce_settings.test", "enabled"),
+					resource.TestCheckResourceAttrSet("data.sendgrid_bounce_settings.test", "soft_bounces"),
+					resource.TestCheckResourceAttrSet("data.sendgrid_bounce_settings.test", "hard_bounces"),
+				),
+			},
+		},
+	})
+}
+
+const testAccBounceSettingsDataSourceConfig = `
+data "sendgrid_bounce_settings" "test" {}
+`